Fixing remove methods
Adding unit tests for remove methods.  Also adding workaround
for when we remove all the particles or bonds from our particle
container as pytables throws an exception (NotImplementedError)
 if we remove all rows from a table ("You are trying to
delete all the rows in a table" ... "This is not
supported right now due to limitations on the underlying HDF5 library.
Sorry!"). So in such cases, instead of removing the last row
of the table, we remove the old table and create a new one.